[#1308] Move label-specific grouping tests.
